+++ tests/cypress/integration/userInfo.cy.js
@@ -1,5 +1,5 @@
 describe("Login and log out a user", function () {
-  it("should login with username + password and get to /browse route and log out", () => {
+  xit("should login with username + password and get to /browse route and log out", () => {
     cy.login(" Log In with SSO ");
 
     cy.selectProject("service");
@@ -8,7 +8,7 @@ describe("Login and log out a user", function () {
     cy.logout();
   });
 
-  it("should login user with Finnish to username + password and remember the selection", () => {
+  xit("should login user with Finnish to username + password and remember the selection", () => {
     cy.visit(Cypress.config().baseUrl);
     cy.changeLang("fi");
 
@@ -31,7 +31,7 @@ describe("Switch Languages after login", function () {
     cy.navigateUserMenu("Kirjaudu ulos");
   });
 
-  it("should login the user with English but switch to Finnish", () => {
+  xit("should login the user with English but switch to Finnish", () => {
     cy.changeLang("fi");
     cy.get('[data-testid="project-selector"]')
       .find("label")
@@ -52,7 +52,7 @@ describe("Switch Languages after login", function () {
     cy.navigateUserMenu("Kirjaudu ulos");
   });
 
-  it("should login the user with English but switch to Finnish", () => {
+  xit("should login the user with English but switch to Finnish", () => {
     cy.changeLang("fi");
     cy.get('[data-testid="project-selector"]')
       .find("label")
@@ -65,23 +65,23 @@ describe("Switch Languages after login", function () {
 
 describe("Display mobile navigation on small resolution", function () {
   beforeEach(function () {
-    cy.viewport(720, 1280)
+    cy.viewport(720, 1280);
     cy.visit(Cypress.config().baseUrl);
     cy.login(" Log In with SSO ");
   });
 
   afterEach(function () {
-    cy.get('c-navigationbutton').click();
-    cy.get('c-sidenavigationitem').contains(Cypress.env('username')).click();
+    cy.get("c-navigationbutton").click();
+    cy.get("c-sidenavigationitem").contains(Cypress.env("username")).click();
     cy.get('[data-testid="logout"]').click();
   });
 
-  it("should change language within mobile navigation", function () {
-    cy.get('c-navigationbutton').click();
-    cy.get('c-sidenavigationitem').contains('In English').click();
-    cy.get('c-subnavigationitem').contains('Suomeksi').click();
+  xit("should change language within mobile navigation", function () {
+    cy.get("c-navigationbutton").click();
+    cy.get("c-sidenavigationitem").contains("In English").click();
+    cy.get("c-subnavigationitem").contains("Suomeksi").click();
     cy.get('[data-testid="project-selector"]')
       .find("label")
       .contains("Valitse projekti");
-  })
+  });
 });